Misc nvim changes.
First, make ~/.config a real directory and put everything in there
that isn't version controlled.  Next, only ln the things that are
in version control to that directory.  This is so that things in the
config directory aren't being .gitignore'd anymore and we can do stuff
like search through them with ripgrep.  Not to mention see when things
have actually changed...

Other misc changes:

1) Disable lasyvim_last_loc so that the cursor position is always at the
   top of a buffer.  I think unpredictability is bad UI.
2) Get rid of our ,ww word wrap toggle since ,uw already does it and
   better.
3) Optimize startup times by:
  a) Disabling persistence.nvim and todo-comments.nvim
  b) Making align.nvim and Comments.nvim only load on CursorMoved
  c) Not loading tokyonight at all, only to use kanagawa in the end.